Adding importers + small touches
By Merlin's beard! We've added the importers, HackerNews API isn't the
greatest but it was still easy to work with, and I'm sure it's easy for
them to maintain `https://github.com/HackerNews/API`. Importers are
pretty straight forward with regards to what they do. Currently we're
using the API Client as maybe at one point I see the importers split
into a separate service